Move mission steps to a dedicated package
