+++ mars-sim-core/pom.xml
@@ -12,25 +12,10 @@
 	<name>mars-sim-core</name>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<twelvemonkeys.version>3.12.0</twelvemonkeys.version>
 		<jogamp.version>2.5.0</jogamp.version>
 	</properties>
 	<dependencies>
-		<dependency>
-		    <groupId>com.twelvemonkeys.imageio</groupId>
-		    <artifactId>imageio-core</artifactId>
-	        <version>${twelvemonkeys.version}</version>
-		</dependency>
-		<dependency>
-		    <groupId>com.twelvemonkeys.imageio</groupId>
-		    <artifactId>imageio-metadata</artifactId>
-	        <version>${twelvemonkeys.version}</version>
-		</dependency>
-	    <dependency>
-	        <groupId>com.twelvemonkeys.imageio</groupId>
-	        <artifactId>imageio-jpeg</artifactId>
-	        <version>${twelvemonkeys.version}</version>
-	    </dependency>
+		
 		<dependency>
 		    <groupId>commons-cli</groupId>
 		    <artifactId>commons-cli</artifactId>
