+++ mars-sim-core/src/main/java/com/mars_sim/core/mission/predefined/TestDriveMission.java
@@ -12,8 +12,8 @@ import java.util.List;
 import com.mars_sim.core.map.location.Coordinates;
 import com.mars_sim.core.map.location.Direction;
 import com.mars_sim.core.mission.MissionStep;
-import com.mars_sim.core.mission.MissionTravelStep;
 import com.mars_sim.core.mission.MissionVehicleProject;
+import com.mars_sim.core.mission.steps.MissionTravelStep;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.mission.MissionType;
 import com.mars_sim.core.person.ai.mission.NavPoint;
